prototype development

+++ feed.go
@@ -6,65 +6,81 @@ import (
 	"io/ioutil"
 	"os"
 	"path"
+	"strings"
 )
 
 func feedFolder(folder string) {
-	fmt.Println("Feeding folder: " + folder)
+	fmt.Println("Feeding: " + folder + "\nStarting..." + "\n-------")
 	files, err := ioutil.ReadDir(folder)
 	if err != nil {
-		panic(err)
+		fmt.Println("Feeding: " + folder + "\nError: " + err.Error() + "\n-------")
+		return
 	}
 	for _, inside := range files {
 		doing := path.Join(folder, inside.Name())
 		if inside.IsDir() {
 			feedFolder(doing)
 		} else {
-			feedFile(doing)
+			waiter.Add(1)
+			go feedFile(doing)
 		}
 	}
-	os.Remove(folder)
+	if clean {
+		os.Remove(folder)
+	}
 }
 
-func feedFile(file string) {
-	fmt.Println("Feeding file: " + file)
-	sts, err := os.Stat(file)
-	if os.IsNotExist(err) {
-		panic("The feed file does not exists: " + file)
-	}
-	if sts.Size() == 0 {
-		fmt.Println("The file is empty: " + file)
+func feedFile(origin string) {
+	defer waiter.Done()
+	fmt.Println("Feeding: " + origin + "\nStarting..." + "\n-------")
+	exType := strings.TrimSpace(strings.ToLower(path.Ext(origin)))
+	if !enabledTypes[exType] {
+		fmt.Println("Feeding: " + origin + "\nError: It's not an enabled type." + "\n-------")
 		return
 	}
-	data, err := ioutil.ReadFile(file)
+	sts, err := os.Stat(origin)
+	if os.IsNotExist(err) {
+		fmt.Println("Feeding: " + origin + "\nError: The file does not exists." + "\n-------")
+		return
+	}
+	if sts.Size() == 0 {
+		fmt.Println("Feeding: " + origin + "\nError: The file is empty." + "\n-------")
+		return
+	}
+	data, err := ioutil.ReadFile(origin)
 	if err != nil {
-		panic(err)
+		fmt.Println("Feeding: " + origin + "\nError: " + err.Error() + "\n-------")
+		return
 	}
 	check := fmt.Sprintf("%x", md5.Sum(data))
-	fmt.Println("File Check Sum: " + check)
 	root := path.Join(body, check[0:2], check[2:4])
-	destiny := path.Join(root, check+path.Ext(file))
-	fmt.Println("Destiny: " + destiny)
+	destiny := path.Join(root, check+path.Ext(origin))
+	fmt.Println("Feeding: " + origin + "\nDestiny: " + destiny + "\n-------")
 	_, err = os.Stat(destiny)
 	if os.IsNotExist(err) {
 		os.MkdirAll(root, os.ModePerm)
-		err = os.Rename(file, destiny)
+		err = os.Rename(origin, destiny)
 		if err != nil {
-			panic(err)
+			fmt.Println("Feeding: " + origin + "\nError: " + err.Error() + "\n-------")
+			return
 		}
 	} else {
-		fmt.Println("This file is already on my belly.")
+		fmt.Println("Feeding: " + origin + "\nError: This file is already on my belly." + "\n-------")
 		return
 	}
+	fmt.Println("Feeding: " + origin + "\nResult: Successfully eaten." + "\n-------")
 }
 
 func doFeed() {
 	sts, err := os.Stat(feed)
 	if os.IsNotExist(err) {
-		panic("The feed path does not exists: " + feed)
+		fmt.Println("Feeding: " + feed + "\nError: The path does not exists." + "\n-------")
+		return
 	}
 	if sts.IsDir() {
 		feedFolder(feed)
 	} else {
-		feedFile(feed)
+		waiter.Add(1)
+		go feedFile(feed)
 	}
 }
